ml-service/requirements.txt
ultralytics
torch
torchvision
opencv-python
fastapi
pillow
numpy
uvicorn
python-multipart
